Don't use >>.grep if .grep is enough (#4)
The use of `>>.grep` in combination with stringification is something that feels strange.  In any case, somehow the recent speedup of `>>.` (aka `nodemap`) something subtly changed so that that expression generates a `Seq` with a lot of empty `Seq`s in them, each of which stringifies to a space, and thus breaks the test.

Since I don't see a need for the `>>.` there, I'd thought it safe to remove.